Add files via upload
I encountered an issue with coursera as I was not able to access all of the tests. I was told to just upload these screenshots. From watching the videos I can see that it is material that I previously seen, thus, I do not think this will be an issue when pursuing the statistical module at Ironhack.